Modularização do sistema de pagamentos e implementação do PIX
- Separação da lógica em módulos: base.py, pix.py e cartao.py
- Novo script pagamentos_teste.py para testar pagamentos via PIX e Cartão
