+++ backend/testes_pagamentos.py
@@ -1,0 +1,57 @@
+# backend/main.py
+from pagamentos.pix         import Pix
+from pagamentos.cartao      import CartaoCredito, CartaoDebito
+from pagamentos.base        import Pedido, Logger, Mensageria
+
+def processar_pix():
+    logger     = Logger()
+    mensageria = Mensageria()
+    pedido     = Pedido("João", 120.50)
+    pix        = Pix(pedido, "meu@email.com", logger, mensageria)
+    pix.processar_pagamento()
+
+def processar_cartao_credito():
+    logger     = Logger()
+    mensageria = Mensageria()
+    pedido     = Pedido("Maria", 200.75)
+    cc         = CartaoCredito(
+        pedido,
+        numero="1234567890123456",
+        titular="Maria Silva",
+        validade="12/26",
+        cvv="123",
+        logger=logger,
+        mensageria=mensageria
+    )
+    cc.processar_pagamento()
+
+def processar_cartao_debito():
+    logger     = Logger()
+    mensageria = Mensageria()
+    pedido     = Pedido("Carlos", 150.00)
+    cd         = CartaoDebito(
+        pedido,
+        numero="6543210987654321",
+        titular="Carlos Souza",
+        validade="11/25",
+        cvv="321",
+        logger=logger,
+        mensageria=mensageria
+    )
+    cd.processar_pagamento()
+
+if __name__ == "__main__":
+    print("Escolha o método de pagamento:")
+    print("1. PIX")
+    print("2. Cartão de Crédito")
+    print("3. Cartão de Débito")
+    escolha = input("Digite 1, 2 ou 3: ")
+
+    if escolha == "1":
+        processar_pix()
+    elif escolha == "2":
+        processar_cartao_credito()
+    elif escolha == "3":
+        processar_cartao_debito()
+    else:
+        print("Opção inválida!")
